+ Adding support for Microsoft Visual Studio Express, MVSE:   - I can build and test most of Draco with MVSE 2010.  I have one     issue to resolve in parser and a more serious issue with     quadrature.  I have not tried to port FortranChecks, cdi_eospac or     device at this time.  All other components build and pass their     tests under MVSE.   - When building unit tests add a post_build command to provide debug     pdb (debug information) files.   - Add support for building against Random123.   - Marked classes and methods with the CPP macro DLL_EXPORT so that     MVSE marks these symbols as public in the generated library.   - Added some MVSE pragma instructions to suppress a couple of     annoying compile warnings (C4251).   - In tstIpcress_Interpreter and I made the regression check more     robust so that values read from a file are compared to gold     standard values as numbers instead of strings.  This addresses an     issue where MVSE prints numbers with the format 1.0e+001 vs. Linux     where we have 1.0e+01.
+ I moved the function tokenize() from Procmon.hh to UnitTest.hh as it
  is a useful for the UnitTest framework. 
+ I provided new accessor functions in UnitTest to provide the path
  and name of the current unit test. 
+ I updated get_filename_component to better support the Windows file
  naming scheme.  This update also provided a new feature, FC_NATIVE,
  that will convert a path between Unix format and Windows format. 
+ In many cases, I updated unit tests to use the UnitTest framework
  provided by Draco.  I also replaced the block of catch statements
  with the CPP macro UT_EPILOG(ut).  This change allows the developer
  to focus on the tests instead of the test driver.  For some
  components, I also eliminated the generation of the test library. 
+ Minor tweaks:
  - Removed some dead code from build system files.
  - Added some double quotes around build system variables to allow
    spaces in paths. 
  - Remove some build system code that was specific to Roadrunner. 
  - Updated copyright and other header information.

+ Updated path handling in unit tests so that absolute paths are used
  via rtt_dsxx::UnitTest::getTestPath().  This eliminates some
  assumptions in unit tests about paths and file locations.  Many
  IDE's like Visual Studio or XCode use a directory structure of the
  form: 
   norms/
      Debug/
      Release/
      test/
         Debug/ 
         Release/
  For these IDE's the tests and input files are in a Configuration
  subdirectory instead of in 'test/' but ctest is still run from
  'test/'  Thus, the input files are no longer located at './'  Using
  a more complete path handling  system resolves the issue. 

refs #135 @9h

+++ src/cdi_analytic/Pseudo_Line_Analytic_Odfmg_Opacity.cc
@@ -244,7 +244,7 @@ Pseudo_Line_Analytic_Odfmg_Opacity::getOpacity(double T,
                         b1 = bands[b+1];
                         // little bit of an elaborate interpolation dance here
                         double f = 0.5*(b0 + b1)*N - 0.5;
-                        unsigned const i = f;
+                        unsigned const i = static_cast<unsigned int>(f);
                         f -= i;
                         Check(i<N);
                         Check(f>=-0.5 && f<=0.5);
